refactor: move the vtbackend cell and line storage into grid/
The buffer and what it is made of: the scrollback ring and its reflow (Grid), a
row in both its trivial and inflated forms (Line, LineSoA), the writer that
decodes UTF-8 and segments graphemes into it (SoAClusterWriter), and the views
onto a single cell (CellProxy, CellUtil).

What is NOT here is as much of the point. The flags enums went to core/ in the
previous commit, because GraphicsAttributes and the whole OSC 133 group read them
and neither has any business depending on the grid. Selection is not here either
-- Selector reaches cells only through its SelectionHelper seam and names nothing
in this directory -- and neither is Viewport, which includes Screen and is a view
onto the terminal rather than onto the grid.
