+++ src/vtbackend/RenderBuffer.hpp
@@ -1,13 +1,13 @@
 // SPDX-License-Identifier: Apache-2.0
 #pragma once
 
-#include <vtbackend/Grid.hpp>
-#include <vtbackend/Line.hpp>
 #include <vtbackend/core/CellFlags.hpp>
 #include <vtbackend/core/Color.hpp>
 #include <vtbackend/core/Image.hpp>
 #include <vtbackend/core/Primitives.hpp>
 #include <vtbackend/core/TextScale.hpp>
+#include <vtbackend/grid/Grid.hpp>
+#include <vtbackend/grid/Line.hpp>
 
 #include <gsl/pointers>
