+++ src/vtbackend/VTWriter.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 #include <vtbackend/VTWriter.hpp>
 
-#include <vtbackend/CellProxy.hpp>
+#include <vtbackend/grid/CellProxy.hpp>
 
 #include <numeric>
