+++ src/vtbackend/grid/CellProxy.hpp
@@ -1,15 +1,15 @@
 // SPDX-License-Identifier: Apache-2.0
 #pragma once
 
-#include <vtbackend/CellUtil.hpp>
-#include <vtbackend/LineSoA.hpp>
-#include <vtbackend/SoAClusterWriter.hpp>
 #include <vtbackend/core/CellFlags.hpp>
 #include <vtbackend/core/Color.hpp>
 #include <vtbackend/core/GraphicsAttributes.hpp>
 #include <vtbackend/core/Hyperlink.hpp>
 #include <vtbackend/core/Image.hpp>
 #include <vtbackend/core/Primitives.hpp>
+#include <vtbackend/grid/CellUtil.hpp>
+#include <vtbackend/grid/LineSoA.hpp>
+#include <vtbackend/grid/SoAClusterWriter.hpp>
 
 #include <libunicode/convert.h>
 #include <libunicode/width.h>
