+++ src/vtbackend/grid/Grid_test.cpp
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/Grid.hpp>
 #include <vtbackend/core/Primitives.hpp>
+#include <vtbackend/grid/Grid.hpp>
 
 #include <crispy/BufferObject.hpp>
