+++ src/vtbackend/grid/LineSoA_test.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/CellProxy.hpp>
-#include <vtbackend/Line.hpp>
-#include <vtbackend/LineSoA.hpp>
+#include <vtbackend/grid/CellProxy.hpp>
+#include <vtbackend/grid/Line.hpp>
+#include <vtbackend/grid/LineSoA.hpp>
 
 #include <catch2/catch_test_macros.hpp>
