+++ src/vtrasterizer/RenderTarget.hpp
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 #pragma once
 
-#include <vtbackend/Grid.hpp> // cell attribs
 #include <vtbackend/core/Color.hpp>
 #include <vtbackend/core/Primitives.hpp>
+#include <vtbackend/grid/Grid.hpp> // cell attribs
 
 #include <vtrasterizer/GridMetrics.hpp>
 #include <vtrasterizer/ImageTextureBackend.hpp>
